+++ database/migrations/2015_12_07_061307_create_forum_covers_table.php
@@ -18,7 +18,7 @@ class CreateForumCoversTable extends Migration
         Schema::create('forum_forum_covers', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->mediumInteger('forum_id')->unsigned()->nullable();
-            $table->mediumInteger('user_id')->unsigned()->nullable();
+            $table->integer('user_id')->unsigned()->nullable();
             $table->string('hash');
             $table->string('ext');
             $table->timestamps();
