+++ database/migrations/2016_04_12_111756_add_last_editor_id_to_beatmap_discussion_posts.php
@@ -15,7 +15,7 @@ class AddLastEditorIdToBeatmapDiscussionPosts extends Migration
     public function up()
     {
         Schema::table('beatmap_discussion_posts', function ($table) {
-            $table->mediumInteger('last_editor_id')->unsigned()->nullable();
+            $table->integer('last_editor_id')->unsigned()->nullable();
 
             $table->foreign('last_editor_id')
                 ->references('user_id')
