+++ database/migrations/2016_07_29_024320_create_contest_tables.php
@@ -62,7 +62,7 @@ class CreateContestTables extends Migration
                 ->on('contests')
                 ->onDelete('restrict');
 
-            $table->mediumInteger('user_id')->unsigned();
+            $table->integer('user_id')->unsigned();
             $table->foreign('user_id')
                 ->references('user_id')
                 ->on('phpbb_users')
